Reimplement content (transcluded features) for collaborationhubs
This implements model-agnostic transclusion, makes everything a whole lot
neater than it was, and generally raises far more questions than it
answers.

I need to file a bunch of bugs to follow up on this, and it's probably
full of holes to boot. Something about how no styles are being applied
from collaborationlists (probably the same for sub-hubs, though
haven't tested, but the collaborationlist styles are something we
want here), instantcommons isn't working (possibly unrelated),
galleries weren't working (haven't tested them now), we need it to
only include part of the page, not all of it (and probably respect
noincludes etc)...

bug: T136785
Change-Id: Ifab7ea02b22e720435c6cc7f77f8d1d2937df654