Adding membership-related parameters to the CollaborationListContent validator.
The validator implementation forgot to account for the ismemberlist boolean that, if flipped on, causes the list to be rendered totally different (once it is actually implemented per T141018). The memberoptions array should also be considered valid for the time being, since it is a part of the schema, even though it is not being used for anything yet.

Also made some other changes for consistency.

Change-Id: Icfb213bbe897d563f7bbd8d585ea3c69f30b22c9