Made a separate ToC and Icon classes for collaborationhubs and collaborationkit in general, respectively.
Icons are generally static and called to render from the canned or onwiki
inmages.
ToC is called as part of the Hub rendering process.

Icons probably need to be called more places, or stuff. Also we may need a js
version, given that it's js anyway. But this is there, now.

bug: T140170
Change-Id: I3126a963c603c2bd4ca38b0552eb52e14697eaa2