Add __NOCOLLABORATIONHUBTOC__ magic word
This change also causes the TOC to only be displayed when
displaying an actual page. It won't display any more on
history pages, pages that don't exist, etc. Hopefully that is
ok.

The way this is implemented is kind of hacky. Its the best I can
come up with at the moment :s

Bug: T141009
Change-Id: I33fcc3129e175607c72199b017dd2b613d1fa69f