Removing edit section links and re-implementing some as OOjs UI frameless buttons
The problem with the stock edit section links is that where VisualEditor is installed, the message, regardless of its content, is replaced with 'edit source'. Rather than fix this, we sidestep it by just using OOjs UI frameless buttons. However, due to other design issues, I have not re-implemented all the buttons; just the "view" button for each feature and an "edit" button by announcements.

Bug: T148877
Change-Id: I2d700d1a423892a9643cbc909767e8d610cf7979