Style collaborationlist columns
Multicolumn lists now show up side-by side, with minimum and maximim
widths, using flexbox to handle sizing and fill and overlow.
Columns are styled as blocks using theme colour, if any, or defaulting
to a nice grey if not. Almost made it default to a @default blue, but
then changed my mind.

Non-multicolumn lists remain the same.

Also added a NOTOC to the collaborationlist wikitext to prevent a toc
from randomly showing up when too many headers exist. Possibly not the
best approach.

Bug: T149035
Change-Id: I736137663c5bdf27a5aa8d04c1bb2c89f785bd4f