Style custom edit forms a bit
Doesn't style around the js tools, but this layout should act as a
fallback either way.

Also does some magic to use the hub colour, if it's a hub, for the
borders.
Used zz as a prefix for the module name to ensure it loads last to
avoid need for stupid css overrides.

Put back @default.

Bug: T148943
Bug: T149031
Change-Id: I1593f6dc199e0f0f639c891436429443a6d1519d