One-click project join from within the hub.
Clicking the blue "Join" button on a hub will now have you join as a member just in one click. You are then taken to the member list.

Once you join as a member, the blue "Join" member disappears.

Going to the member page's action=edit remains as a no-JS fallback

Change-Id: Iebbf8d2c1c06912f73a1f52f74c34fb42e78d9c3