Preventing hub transclusion from causing exception.
When a CollaborationHubContent page is transcluded, it will return the hub image and intro. It can't do more because much of CollaborationHubContent relies on fillParserOutput (i.e. straight HTML) rather than something like wikitext that can be fed into the parser. Would be worth investigating to see if this can be fixed.

Bug: T157228
Change-Id: Id0d27a755e6cec02ec0efaf8fc99df54139724e6