Namespace selector for CreateCollaborationHub
Page is created in the namespace selected from the dropdown. I gutted the validation callback because I couldn't figure out how to get it to work with two pieces of input, so I just do title validation at a later stage.

The final product should *not* look like this. I know Isarra is working on the styling so I am deferring to her.

Bug: T155859
Change-Id: Ided3791360ebb127d96c380be015090e449a2115