Make edit form visually distinct from preview content regardless of user settings
User may or may not have an edit toolbar; top border should
render regardless.
User may have preview above or below; form should still be
visually distinct if preview is above, so entire thing is
now wrapped in a coloured box like the editoptions.

I don't like this. But there's no meaningful boundary
between the preview and the editform when the preview is
before the editform otherwise. And this is the default
setting.

Change-Id: I8a3f202964964901e9f8b49200beac2920930cd7