Remove JS functionality for logged-out users for join button.
The project join button assumes the existence of a user account. Logged
out users lack this, so this removes the "join" button from the member
list itself. The button is still technically there on the hub page, but
the JavaScript functionality is removed, so logged out users can still
technically join projects (to the extent they're already allowed to join
WikiProjects).

Change-Id: Iadc48eca8f9ed65d32578148b0cd54903c5d4e7f