Fixed everything to do with transcluding lists.
1. Adding columns= as a parameter for {{#transcludelist}}

Behaves in the same way as tags= but for columns. The exception was
caused by "columns" being a valid option but not having its own
special handling. While this will technically work, in the docs we
should continue to prefer |column=|column= because that way we avoid
the follies of having the delimiter be something that could validly
show up as a column name ("Art+Feminism," for example).

2. offset and iconWidth

Solved by casting the values to (int). Also changed how icons are
processed in general. This meant getting rid of the height/width rule
for item icons and relying on element styling, which is lousy but
that's the price we pay for arbitrary icon size.

Also, offset processing was only applying itself to the first column
so I fixed that.

3. includeDesc and showColumnHeaders

Check if the inputted false is 'false', 'no', or 0, and declare those
parameters as false if so. Fun fact: (bool)'false' casts to true!

I am sorry for this all being one commit, but these were all small,
interconnected changes.

Bug: T159398
Bug: T159399
Bug: T159400
Change-Id: Ief0e64b703218a74580fa934a00fe94219c9059e