Change JS list editor to use numeric indices instead of titles.
A page title could be included on a list multiple times, and this
causes issues with the JS list editor. This change assigns "item
IDs" to each list item, based on the column and item index, and
changes all list editor behavior to use those instead of the page
title.

Change-Id: I01af89264c59f943274c422b630b1f9fb09622ca