JSDoc: switch from JSDuck to JSDoc
IMPORTANT: Probably switches doc.wikimedia.org from the existing
JSDuck site to JSDoc either immediately or when the next train
runs.

- Switch from JSDuck to JSDoc. Had to because @memberof causes
JSDuck to fail CI.
- Fix missing class methods in /docs/ by adding @memberof

I added # to the end of each @memberof statement
to mark it as an instance instead of static. If this is
not correct, let me know. Adding # got rid of an eslint
warning.

Bug: T366789
Change-Id: I1c2e50f9fb61d8ef9eee7a1dbb781af2a3ed874d